% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find_praat}
\alias{find_praat}
\title{praatsyllable: R Package for Syllable Detection using Praat}
\usage{
find_praat()
}
\description{
Integrates the Praat Syllable Nuclei script for detecting syllables in audio files
}
