% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_syllable_count}
\alias{get_syllable_count}
\title{Read syllable count from TextGrid}
\usage{
get_syllable_count(textgrid_path)
}
\arguments{
\item{textgrid_path}{Path to TextGrid file}
}
\value{
Number of syllables detected
}
\description{
Read syllable count from TextGrid
}
