% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_syllables_ps}
\alias{get_syllables_ps}
\title{Get Syllables Per Second}
\usage{
get_syllables_ps(
  wav_file,
  ignorance_level = 0,
  min_dip = 2,
  min_pitch = 100,
  time_step = 0,
  threshold = -45,
  min_silent_int = 0.5,
  min_sound_int = 0.1,
  praat_path = NULL,
  keep_intermediates = FALSE
)
}
\arguments{
\item{wav_file}{Path to input WAV file}

\item{ignorance_level}{Ignorance level/Intensity median (dB). Default 0 for unfiltered sounds, 2 for filtered}

\item{min_dip}{Minimum dip between peaks (dB). Default 2 for unfiltered sounds, 4 for filtered}

\item{min_pitch}{Minimum pitch (Hz)}

\item{time_step}{Time step (s)}

\item{threshold}{silence threshold, default is -45}

\item{min_silent_int}{Minimum silent interval (s)}

\item{min_sound_int}{Minimum sounding interval (s)}

\item{praat_path}{Optional path to Praat executable}

\item{keep_intermediates}{Keep the tmp folder with the textgrids? TRUE is kept, FALSE deletes the intermediate files}
}
\value{
Turn lengths
}
\description{
Get Syllables Per Second
}
