% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_turn_duration}
\alias{get_turn_duration}
\title{Get Speaking Turn Lengths}
\usage{
get_turn_duration(
  wav_file,
  min_pitch = 100,
  time_step = 0,
  threshold = -45,
  min_silent_int = 0.5,
  min_sound_int = 0.1
)
}
\arguments{
\item{wav_file}{Path to input WAV file}

\item{min_pitch}{Minimum pitch (Hz)}

\item{time_step}{Time step (s)}

\item{threshold}{silence threshold, default is -45}

\item{min_silent_int}{Minimum silent interval (s)}

\item{min_sound_int}{Minimum sounding interval (s)}
}
\value{
Turn lengths
}
\description{
Get Speaking Turn Lengths
}
