% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{split_channels}
\alias{get_boundaries}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{wav2textgrid}{\code{\link[wav2textgrid]{get_boundaries}}, \code{\link[wav2textgrid]{split_channels}}}
}}

